requirements.txt
asgiref==3.6.0
Django==4.2.1
django-filter==23.2
django-request-logging==0.7.5
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
